refactor: `OauthPlatforms` 객체 enum 으로 수정
