+++ frontend/src/apis/login.ts
@@ -1,9 +1,8 @@
-import { getOauthURL } from 'constants/apis/url';
+import { PostOauthLoginRequest } from 'types/apis/login';
 import { http } from './fetch';
-import { OauthPlatform, PostOauthLoginRequest } from 'types/apis/login';
-import { getRedirectURL } from 'constants/apis/oauth';
+import { OauthPlatforms, getOauthURL, getRedirectURL } from 'constants/components/oauth';
 
-export const getRedirection = (platform: OauthPlatform) =>
+export const getRedirection = (platform: OauthPlatforms) =>
   http.get(`${getOauthURL(platform)}?redirect_uri=${getRedirectURL(platform)}`);
 
 export const postOauthLogin = ({ platform, body }: PostOauthLoginRequest) =>
