+++ frontend/src/mocks/handlers/login.ts
@@ -1,15 +1,17 @@
-import { getRedirectURL } from 'constants/apis/oauth';
-import { getOauthURL } from 'constants/apis/url';
+import { OauthPlatforms, getOauthURL, getRedirectURL } from 'constants/components/oauth';
 import { rest } from 'msw';
 
 export const loginHandlers = [
   // 카카오 로그인 페이지로 리다이렉트 요청: GET
-  rest.get(`${getOauthURL('kakao')}?redirect_uri=${getRedirectURL('kakao')}`, (_, res, ctx) => {
-    return res(ctx.delay(300), ctx.status(302));
-  }),
+  rest.get(
+    `${getOauthURL(OauthPlatforms.kakao)}?redirect_uri=${getRedirectURL(OauthPlatforms.kakao)}`,
+    (_, res, ctx) => {
+      return res(ctx.delay(300), ctx.status(302));
+    },
+  ),
 
   // 카카오 로그인/회원가입: POST
-  rest.post(getOauthURL('kakao'), (_, res, ctx) => {
+  rest.post(getOauthURL(OauthPlatforms.kakao), (_, res, ctx) => {
     return res(ctx.delay(300), ctx.status(200));
   }),
 ];
