+++ frontend/src/pages/OauthPage/OauthPage.tsx
@@ -1,10 +1,9 @@
 import { useMutation } from '@tanstack/react-query';
 import { postOauthLogin } from 'apis/login';
-import { Platforms, getRedirectURL } from 'constants/apis/oauth';
+import { OauthPlatforms, getRedirectURL } from 'constants/components/oauth';
 import { usePageNavigate } from 'hooks/usePageNavigate';
 import { useEffect } from 'react';
 import { useLocation, useSearchParams } from 'react-router-dom';
-import { OauthPlatform } from 'types/apis/login';
 
 const OauthPage = () => {
   const { goHomePage } = usePageNavigate();
@@ -23,8 +22,8 @@ const OauthPage = () => {
     const code = searchParams.get('code');
     const platform = location.pathname.split('/').pop();
 
-    const isOauthPlatform = (platform: string | undefined): platform is OauthPlatform => {
-      return platform ? platform in Platforms : false;
+    const isOauthPlatform = (platform: string | undefined): platform is OauthPlatforms => {
+      return platform ? platform in OauthPlatforms : false;
     };
 
     if (!isOauthPlatform(platform) || !code) {
